Tests for editing past bookings in admin and disclaimer re-signing
