+++ booking/tests/test_block_views.py
@@ -1,4 +1,5 @@
 from datetime import datetime, timedelta
+from mock import patch
 from model_mommy import mommy
 
 from django.conf import settings
@@ -400,6 +401,30 @@ class BlockListViewTests(TestSetupMixin, TestCase):
         )
         self.assertNotIn(
             'Please complete a disclaimer form before buying a block.',
+            format_content(resp.rendered_content)
+        )
+
+        # expired disclaimer
+        user_expired_disclaimer = User.objects.create_user(
+            username='test_expired', email='test@test.com', password='test'
+        )
+        self.client.login(
+            username=user_expired_disclaimer.username, password='test'
+        )
+        field = OnlineDisclaimer._meta.get_field('date')
+        mock_now = lambda: datetime(2015, 2, 10, 19, 0, tzinfo=timezone.utc)
+        with patch.object(field, 'default', new=mock_now):
+            disclaimer = mommy.make_recipe(
+               'booking.online_disclaimer', user=user_expired_disclaimer
+            )
+        self.assertFalse(disclaimer.is_active)
+        resp = self.client.get(reverse('booking:block_list'))
+        self.assertNotIn(
+            'Get a new block!', format_content(resp.rendered_content)
+        )
+        self.assertIn(
+            'Your disclaimer has expired. Please review and confirm your '
+            'information before buying a block.',
             format_content(resp.rendered_content)
         )
